rename densities -> logdensities
