+++ gpflow/likelihoods.py
@@ -20,7 +20,7 @@ import tensorflow as tf
 import numpy as np
 
 from . import settings
-from . import densities
+from . import logdensities
 from . import transforms
 
 from .decors import params_as_tensors
@@ -148,7 +148,7 @@ class Gaussian(Likelihood):
 
     @params_as_tensors
     def logp(self, F, Y):
-        return densities.gaussian(F, Y, self.variance)
+        return logdensities.gaussian(F, Y, self.variance)
 
     @params_as_tensors
     def conditional_mean(self, F):  # pylint: disable=R0201
@@ -164,7 +164,7 @@ class Gaussian(Likelihood):
 
     @params_as_tensors
     def predict_density(self, Fmu, Fvar, Y):
-        return densities.gaussian(Fmu, Y, Fvar + self.variance)
+        return logdensities.gaussian(Fmu, Y, Fvar + self.variance)
 
     @params_as_tensors
     def variational_expectations(self, Fmu, Fvar, Y):
@@ -193,7 +193,7 @@ class Poisson(Likelihood):
         self.binsize = np.double(binsize)
 
     def logp(self, F, Y):
-        return densities.poisson(self.invlink(F) * self.binsize, Y)
+        return logdensities.poisson(self.invlink(F) * self.binsize, Y)
 
     def conditional_variance(self, F):
         return self.invlink(F) * self.binsize
@@ -213,7 +213,7 @@ class Exponential(Likelihood):
         self.invlink = invlink
 
     def logp(self, F, Y):
-        return densities.exponential(self.invlink(F), Y)
+        return logdensities.exponential(self.invlink(F), Y)
 
     def conditional_mean(self, F):
         return self.invlink(F)
@@ -235,7 +235,7 @@ class StudentT(Likelihood):
 
     @params_as_tensors
     def logp(self, F, Y):
-        return densities.student_t(Y, F, self.scale, self.deg_free)
+        return logdensities.student_t(Y, F, self.scale, self.deg_free)
 
     @params_as_tensors
     def conditional_mean(self, F):
@@ -256,7 +256,7 @@ class Bernoulli(Likelihood):
         self.invlink = invlink
 
     def logp(self, F, Y):
-        return densities.bernoulli(self.invlink(F), Y)
+        return logdensities.bernoulli(self.invlink(F), Y)
 
     def predict_mean_and_var(self, Fmu, Fvar):
         if self.invlink is probit:
@@ -268,7 +268,7 @@ class Bernoulli(Likelihood):
 
     def predict_density(self, Fmu, Fvar, Y):
         p = self.predict_mean_and_var(Fmu, Fvar)[0]
-        return densities.bernoulli(p, Y)
+        return logdensities.bernoulli(p, Y)
 
     def conditional_mean(self, F):
         return self.invlink(F)
@@ -290,7 +290,7 @@ class Gamma(Likelihood):
 
     @params_as_tensors
     def logp(self, F, Y):
-        return densities.gamma(self.shape, self.invlink(F), Y)
+        return logdensities.gamma(self.shape, self.invlink(F), Y)
 
     @params_as_tensors
     def conditional_mean(self, F):
@@ -337,7 +337,7 @@ class Beta(Likelihood):
         mean = self.invlink(F)
         alpha = mean * self.scale
         beta = self.scale - alpha
-        return densities.beta(alpha, beta, Y)
+        return logdensities.beta(alpha, beta, Y)
 
     @params_as_tensors
     def conditional_mean(self, F):
